Remove the initContainers block from mongo-deployment
Close #2.

As mentioned in the issue thread and surrounding discussion, the initContainers
block was causing trouble as the pod was starting up and causing the basic
wrapper script to fail. This is something that has *already* been fixed by
Mojib. Removing the block here in order to reduce noise in the repository.