Beta test: Try fix for api changes
This commit includes a simple fix for the api incompatibility issues that we
were facing. As described by issue #6.

The problem seems to have been that the code used to set up the database:
specifically the function used to connect, no longer returns a db instance as
expected. Instead it returns a `client` instance. This holds a `db`function,
which can be called while passing the dbname in order to retrieve the desired
object. This commit proposes to set up the test environment to explore the solution.