Add files for localdev environment
This centers around the need to include scripts for running the
application locally. This commit adds the ability to run the application
locally using docker-compose.

All scripts are housed in a newly created `localdev` subdirectory. We
also include a number of scripts for use in the exercises outlined in
the documentation. The documentation will follow in a seperate commit.
A top level Makefile is included to handle some setup for the exercises.