Simplify the release of helm charts
Use a simplified workflow to publish charts using github pages.
We're also removing the github workflow that was set up to automate the
release of charts. I had not fully understood how the helm package index
was going to be hosted, so actually the workflow would probably be
adequate now that github pages have been set up. Anyway, this is a fix
by simplification, and closes #16.

Signed-off-by: Adam McCartney <adam@mur.at>