Add docs on contributing
Single file containing a bunch of hyperlinks that attempt to give a
brief overview of the state of contributing to this repository. The
opinions found within this document are common sense suggestions, subject
to the creep and petty pace of time and not intended as some set of
religious dogmas to be tattood onto your neck or other conspicuous body
parts of choice.

Signed-off-by: Adam McCartney <adam@mur.at>