Test custom var sha_short
After being dismayed as usual by the round-aboutiness of CI pipelines, we wish
to set a custom short sha variable that will let us create the image tag
that lets us create a usable reference for our vulnerability scan.

After a bit of fudging about with the hotfix/sec-trivy branch, this commit seems
to stabilize a custom variable that we wish to use to create the image
ref. So at the risk of redundancy and repetition, we submit this humble
commit to be merged.

Signed-off-by: Adam McCartney <adam@mur.at>