Cleanup workflow for trivy scanner
There is a quirk in the way that Github uses the ${{ github.repository }}
environment variable to push to the registry. This variable consists of
camel case characters, which causes problems for trivy if we try to use
it in a later stage. This commit tries to eliminate the problem by hard
coding lowercase values to use in the trivy step. Aaaah, hardcoding.
It also bumps the version of ubuntu used in the pipelines.

This version actually removes our hard to achieve short form sha,
replacing it with the latest tag. This way we can be sure to scan the
most recently pushed image.

Signed-off-by: Adam McCartney <adam@mur.at>